+++ device-management-application/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Gitlab CI with clang format job along with sw-platform regressions
 - Added -v (version) argument to get DM application version
 - Clang formatter rules.
+- Added support to extract SP and MM stats trace buffer
 ### Changed
 - Changed back the sw-platform branch for CI back to develop/system-sw
 - Ran clang formatter.